requirements.txt
numpy
PyYAML
soundfile
pylint
